[hotfix][hive] move hive metastore client and shim related classes to package 'org.apache.flink.table.catalog.hive.client'
